scheduler/pyproject.toml
[tool.poetry]
name = "scheduler"
version = "0.1.0"
description = "Handles creation and running of models and trainers for maestro"
authors = ["Sean Perry <sean.hyatt.perry@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
kubernetes = "^30.1.0"
python-socketio = "^5.11.3"
eventlet = "^0.36.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
